Fix remote console for latest 2.0.0 LXCA builds
The code was only recognizing application/json as a valid json
response. We needed to add a broader check to also accept
application/com.lenovo.lxca-v2.0.0+json, since this Content-Type is
return for the latest 2.0.0 LXCA builds.
Makes it broader enough to accept any content type that begins with
'application/' and ends with 'json', so that it is more resilient to
possible LXCA changes in the future.